Add vectorial form

+++ AITrainer/app/src/main/java/com/johndev/aitrainer/Models/ChargeData.kt
@@ -1,3 +1,3 @@
 package com.johndev.aitrainer.Models
 
-data class ChargeData(val id: Int, val w: Float, val jw: Float, val costo: Float = 0.0f)
+data class ChargeData(val id: Int, val w: Double, val jw: Double, val costo: Double = 0.0)
