+++ AITrainer/app/src/main/java/com/johndev/aitrainer/Models/ResultsAuto.kt
@@ -6,8 +6,8 @@ data class ResultsAuto(
     @SerializedName("ID")
     val ID: Int,
     @SerializedName("W0")
-    val W0: Float,
+    val W0: Double,
     @SerializedName("W1")
-    val W1: Float,
+    val W1: Double,
     @SerializedName("J")
-    val J: Float)
+    val J: Double)
